added sonarqube config
