+++ package.json
@@ -39,6 +39,7 @@
         "jest-expo": "~40.0.0",
         "prettier": "2.2.1",
         "pretty-quick": "^3.1.0",
+        "sonarqube-scanner": "^2.8.0",
         "typescript": "~4.0.0",
         "typescript-logging": "^1.0.0"
     },
