+++ sonar-project.properties
@@ -1,0 +1,25 @@
+sonar.projectKey=Frontend
+sonar.projectName=Frontend
+sonar.projectVersion=1.0
+
+sonar.sources=src/main/
+sonar.host.url=http://localhost:9001
+
+sonar.tests=src/test/
+sonar.coverage.jacoco.xmlReportPaths=build/reports/jacoco/test/jacocoTestReport.xml
+sonar.java.codeCoveragePlugin=jacoco
+sonar.junit.reportPaths=build/test-results/test,build/test-results/integrationTest
+
+sonar.sourceEncoding=UTF-8
+sonar.exclusions=src/main/webapp/content/**/*.*, src/main/webapp/i18n/*.js, build/resources/main/static/**/*.*
+
+sonar.issue.ignore.multicriteria=S3437,S4684,UndocumentedApi
+# Rule https://sonarcloud.io/coding_rules?open=squid%3AS3437&rule_key=squid%3AS3437 is ignored, as a JPA-managed field cannot be transient
+sonar.issue.ignore.multicriteria.S3437.resourceKey=src/main/java/**/*
+sonar.issue.ignore.multicriteria.S3437.ruleKey=squid:S3437
+# Rule https://sonarcloud.io/coding_rules?open=squid%3AUndocumentedApi&rule_key=squid%3AUndocumentedApi is ignored, as we want to follow "clean code" guidelines and classes, methods and arguments names should be self-explanatory
+sonar.issue.ignore.multicriteria.UndocumentedApi.resourceKey=src/main/java/**/*
+sonar.issue.ignore.multicriteria.UndocumentedApi.ruleKey=squid:UndocumentedApi
+# Rule https://sonarcloud.io/coding_rules?open=java%3AS4684&rule_key=java%3AS4684
+sonar.issue.ignore.multicriteria.S4684.resourceKey=src/main/java/**/*
+sonar.issue.ignore.multicriteria.S4684.ruleKey=java:S4684
